Add files via upload
This model is trained on using bert for intent recognition. Using pretrained bert, a dropout layer of 50 % probability, a fully connected layer of 768 units with hyperbolic tangent activation, and then another fully connected layer with the same amount of units as classes. Preprocessing is pretty simple just worrying about adding starting and ending tokens to each tokenized sequence, and then converting those tokens to ids, which we then feed into BERT which creates the embeddings for each id. We train on approx 14,000 data points.