Add files via upload
Tensorflow implementation of using bert and fine tuning it for QA model. The specific inputs we setup for our neural network depend on what the specific model takes in as they all dont take in the same parameters. One key thing I was confused about when trying to make the model was if SparseCategoricalCrossentropy is a valid metric and it is. This also made me realize I need to upgrade my GPU so i can fine tune all the parameters rather then have to freeze the BERT layer for the sake of avoiding crashing. This process is pretty simple, take into account what parameters our specific model needs, here it needs word token ids, the mask corresponding to which indices to perform attention on, and the input type ids. We are able to implement this pretty easily, then we put these all into a list after converting them into a numpy array. Which we can then input into our model(after building). The output is a start and end index probability output given by softmax. Overall this cleared up a lot of the engineering specific confusions I had on question answering models and how to build them with tensorflow, cant wait to get a better GPU so I can deploy full fine tuning!