Add files via upload
Built a 'weird' Transformer model. Will next be working on building different transformer variations with different end tasks in mind. Here the end task in mind when building was for machine translation as the final dense layer had units equivalent to that of the target vocabulary size. However next I will switch it up when I build another variant. I am also going to do a full project involving preprocessing as well very soon. This is just to show some comfortability with TensorFlow.