Add files via upload
FULL PREPROCESSING OF ENGLISH/GERMAN DATA. THEN BUILT TRANSFORMER MODEL FROM SCRATCH AND USED TENSORFLOWS GRADIENTTAPE() TO OPTIMIZE THE MODEL. This took me a bit to do. However this is just the first draft, I am going to be editing this transformer heavily. Additionally I aim that by later tonight I can upload a file of me using this transformer for predictions. This is exciting stuff!